pyproject.toml
[tool.poetry]
name = "muzero"
version = "0.1.0"
description = ""
authors = ["Fraser Mince <frasermince@gmail.com>"]


# jax = {extras = ["tpu"], version = "^0.3.13"}
# poetry run pip install "jax[tpu]>=0.3.13" -f https://storage.googleapis.com/jax-releases/libtpu_releases.html
[tool.poetry.dependencies]
python = ">=3.9, <3.11"
numpy = "^1.20.2"
tensorboard = "^2.4.0"
PyVirtualDisplay = "^2.1"
wandb = "^0.10.27"
sklearn = "^0.0"
ipdb = "^0.13.7"
ale-py = "0.7.5"
opencv-python = "^4.5.5.64"
gym = {extras = ["atari", "accept-rom-license"], version = "^0.23.0"}
packaging = "^20.9"
jax = "^0.3.13"
jaxlib = "^0.3.10"
flax = "^0.5.2"
chex = "^0.1.3"
jaxline = "^0.0.5"
absl-py = "^1.1.0"
tensorflow = "^2.9.1"
ml-collections = "^0.1.1"
ray = "^2.0.0"
fabric = "^2.7.1"
google-api-python-client = "^2.58.0"
cloud-tpu-profiler = ">=2.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
